fix: move public permission service methods to PermissionService
